+++ www/config/build.php
@@ -1,17 +1,17 @@
 <?php
-    // #################################################################################
-    // VERSION INFORMATION - you shouldnt touch those
-    // #################################################################################
+    // #########################################################################
+    // VERSION INFORMATION - dont' touch those
+    // #########################################################################
     //
-    // name of your notes system
-    $m_name	= "monoto";
-
     // version / release
     $m_version = "4.0.0 RC1"; // version
-    $m_date	= "20190305"; // build-date
-    $m_release = false; // true for officials builds
+    $m_date	= "20190307"; // build-date
+    $m_stableRelease = false; // true for officials builds
 
-    // github links
+    // #########################################################################
+    // Github Links
+    // #########################################################################
+    //
     $m_githubProjectLink = "https://github.com/yafp/monoto";
     $m_githubIssueLink = "https://github.com/yafp/monoto/issues";
     $m_githubChangelogLink = "https://github.com/yafp/monoto/blob/master/docs/CHANGELOG.md";