+++ www/inc/checkSession.php
@@ -1,13 +1,19 @@
 <?php
+
 // -----------------------------------------------------------------------------
 // checkSession.php
 // Checks for valid sesstion and redirects back to login page if needed
 // -----------------------------------------------------------------------------
 
+// prevent direct call of this script
+if (strpos($_SERVER['SCRIPT_FILENAME'], 'checkSession.php') !== false)
+{
+    header('Location: ../index.php'); // back to login page
+    die();
+}
 
 // Start a session if it isnt already started
 //
-//session_start();
 if ( session_status() == PHP_SESSION_NONE)
 {
     session_start();
@@ -19,4 +25,6 @@ if ( $_SESSION[ 'monoto' ][ 'valid' ] != 1 )
     header('Location: index.php'); // back to login page
     die();
 }
+
+
 ?>