+++ www/inc/genericIncludes.php
@@ -1,22 +1,31 @@
 <?php
+
 // -----------------------------------------------------------------------------
 // genericIncludes.php
 // Gets included in main files, includes most relevant files itself
 // -----------------------------------------------------------------------------
 
-    // HTML Head
-    include 'inc/genericHeaderHTML.php';
+// prevent direct call of this script
+if (strpos($_SERVER['SCRIPT_FILENAME'], 'genericIncludes.php') !== false)
+{
+    header('Location: ../index.php'); // back to login page
+    die();
+}
 
-    // CSS
-    include 'inc/genericHeaderCSS.php';
 
-    // JS
-    include 'inc/genericHeaderJS.php';
+// HTML Head
+include 'inc/genericHeaderHTML.php';
 
-    // specials
-    require 'inc/helperFunctions.php'; // logging, translation etc
-    require 'inc/db.php'; // connect to db
-    require 'config/config.php'; // some settings
-    require 'config/build.php'; // version informations
+// CSS
+include 'inc/genericHeaderCSS.php';
+
+// JS
+include 'inc/genericHeaderJS.php';
+
+// specials
+require 'inc/helperFunctions.php'; // logging, translation etc
+require 'inc/db.php'; // connect to db
+require 'config/config.php'; // some settings
+require 'config/build.php'; // version informations
 
 ?>