+++ www/inc/noteGetAllNotes.php
@@ -4,6 +4,7 @@
 // Fetches all notes from a single user - server-side processing
 // -----------------------------------------------------------------------------
 
+
 /*
  * DataTables example server-side processing script.
  *