+++ www/inc/noteUpdate.php
@@ -3,14 +3,24 @@
 // noteUpdate.php
 // used to update an existing note from n.php
 // -----------------------------------------------------------------------------
-header('Content-type: text/xml');
+
+// prevent direct call of this script
+if (strpos($_SERVER['SCRIPT_FILENAME'], 'noteUpdate.php') !== false)
+{
+    header('Location: ../index.php'); // back to login page
+    die();
+}
+
+
+//header('Content-type: text/xml');
+header('Content-type: application/xml');
 
 session_start();
 
 if ( $_SESSION[ 'monoto' ][ 'valid' ] == 1 ) // check if the user-session is valid or not
 {
     // note id
-    $modifiedNoteID = filter_input( INPUT_POST, "modifiedNoteID", FILTER_SANITIZE_STRING );
+    $noteID = filter_input( INPUT_POST, "modifiedNoteID", FILTER_SANITIZE_STRING );
 
     // note title
     $modifiedNoteTitle = filter_input( INPUT_POST, "modifiedNoteTitle", FILTER_SANITIZE_STRING );
@@ -37,17 +47,26 @@ if ( $_SESSION[ 'monoto' ][ 'valid' ] == 1 ) // check if the user-session is val
     $owner = $_SESSION[ 'monoto' ][ 'username' ];
 
     // check if the new title is in use already by this user
-    $sql = "SELECT title from m_notes where owner='".$owner."' AND  title='".$modifiedNoteTitle."' and id != ".$modifiedNoteID."' ";
+    $sql = "SELECT title from m_notes where owner='".$owner."' AND  title='".$modifiedNoteTitle."' and id != ".$noteID."' ";
     $result = mysqli_query( $con, $sql );
-    if ( mysqli_num_rows ( $result ) > 0 )
+    // FIXME: check if that really works
+    if ($result == true )
     {
         $current_timestamp = date('Ymd-his');
         $modifiedNoteTitle = $modifiedNoteTitle."___".$current_timestamp;
     }
 
+    /*
+    if ( mysqli_num_rows ( $result ) > 0 )
+    {
+        $current_timestamp = date('Ymd-his');
+        $modifiedNoteTitle = $modifiedNoteTitle."___".$current_timestamp;
+    }
+    */
+
     // update the actual note
     //
-    $sql = "UPDATE m_notes SET title='$modifiedNoteTitle', content='$modifiedNoteContent', save_count='$modifiedNoteCounter' WHERE id='$modifiedNoteID'"; // update m_notes
+    $sql = "UPDATE m_notes SET title='$modifiedNoteTitle', content='$modifiedNoteContent', save_count='$modifiedNoteCounter' WHERE id='$noteID'"; // update m_notes
     $result = mysqli_query($con, $sql);
     if ( !$result )
     {