+++ www/logout.php
@@ -28,7 +28,7 @@
             <?php
                 // update logout conter
                 $con = connectToDatabase();
-                $result = mysqli_query($con, "UPDATE m_users SET logout_counter = logout_counter + 1 WHERE username='".$_SESSION[ 'username' ]."'");
+                $result = mysqli_query($con, "UPDATE m_users SET logout_counter = logout_counter + 1 WHERE username='".$_SESSION[ 'monoto' ][ 'username' ]."'");
 
                 // destroy the user session
                 session_destroy();